+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 5.5.7
 -----------------
+ * Cms content element is_count_views
  * Delete child elements updated
  
 5.5.5